int to sizet for rnd

+++ benchmark/benchmark.cpp
@@ -14,7 +14,8 @@ int slow_get_page(int key) { return key; }
 
 std::vector<int> generate_rnd_keys(size_t n, size_t mod = 100) {
   std::vector<int> keys(n);
-  std::generate(keys.begin(), keys.end(), [&]() { return rand() % mod; });
+  std::generate(keys.begin(), keys.end(),
+                [&]() { return static_cast<size_t>(rand() % mod); });
   return keys;
 }
 
@@ -23,7 +24,7 @@ int main() {
   size_t rnd_mod = 100;
   std::vector<int> keys = generate_rnd_keys(input_sz, rnd_mod);
 
-  std::map<size_t, std::map<std::string, int>> cache_sizes_vs_strategy_hits;
+  std::map<size_t, std::map<std::string, size_t>> cache_sizes_vs_strategy_hits;
   std::cout << "Cache type\tN hits\n";
   for (size_t cache_sz = 10; cache_sz < 101; cache_sz += 10) {
     std::map<std::string, int> cache_hits;
@@ -43,7 +44,7 @@ int main() {
     }
     cache_sizes_vs_strategy_hits[cache_sz] = cache_hits;
 
-  std::cout << "cache size: " << cache_sz << '\n';
+    std::cout << "cache size: " << cache_sz << '\n';
     for (auto [cache_type, nhits] : cache_hits) {
       std::cout << cache_type << "\t" << nhits << '\n';
     }
@@ -53,14 +54,12 @@ int main() {
   // output to a file
   std::ofstream file("benchmark/cache_hits.dat");
   if (file.is_open()) {
-    file << "#Number of keys = " << input_sz << "\n" ;
-    file << "#input keys: uniform distribution [0:" << rnd_mod << ")\n" ;
+    file << "#Number of keys = " << input_sz << "\n";
+    file << "#input keys: uniform distribution [0:" << rnd_mod << ")\n";
     file << "#Cache_size\tLRU\tLFU\tBelady\n";
-    for(auto &[sz, cache_stats]: cache_sizes_vs_strategy_hits){
-      file << sz <<'\t'
-        << cache_stats["LRU"] << '\t'
-        << cache_stats["LFU"] << '\t'
-        << cache_stats["Belady"] << '\n';
+    for (auto &[sz, cache_stats] : cache_sizes_vs_strategy_hits) {
+      file << sz << '\t' << cache_stats["LRU"] << '\t' << cache_stats["LFU"]
+           << '\t' << cache_stats["Belady"] << '\n';
     }
     file.close();
   }
